First version of -scrape command line option (Issue #19). Some scrapers are still not supported. See readme for more details. Also refactored GUI scraping code a bit. Made a few small updates for WebReleaseRenamer list of sites.

+++ JAVMovieScraper/src/moviescraper/doctord/dataitem/Trailer.java
@@ -30,8 +30,12 @@ public class Trailer extends MovieDataItem {
 	}
 	
 	public void writeTrailerToFile(File fileNameToWrite) throws IOException {
-		if(getTrailer() != null && getTrailer().length() > 0)
+		//we don't want to rewrite trailer if the file already exists since that can retrigger a pointlessly long download
+		if(getTrailer() != null && getTrailer().length() > 0 && !fileNameToWrite.exists())
+		{
+			System.out.println("Writing trailer" + this.toString() + " into file " + fileNameToWrite);
 			FileUtils.copyURLToFile(new URL(getTrailer()), fileNameToWrite, connectionTimeout, readTimeout);
+		}
 	}
 	
 	@Override
